+++ CHANGELOG.md
@@ -14,6 +14,7 @@ This document logs the changes per release of TWCManager.
   * (@ngardiner) - Added support for Tesla MFA authentication flows
   * (@MikeBishop) - Improve API error handling, removing transient error delays and replacing with an exponential backoff mechanism to avoid delaying other background tasks.
   * (@MikeBishop) - Added debounce dampening for situations where intermittent consumption spikes / loads cause TWC to start and stop charging frequently.
+  * (@ngardiner) - Add new consumption offset handling which allows for dynamic configuration of offsets in Watts and Amps via web and API
   * Bugfixes
     * (@ngardiner) - Better handling of permissions issues when attempting to save settings.json - alerts user to check file permissions via Web Interface
     * (@ngardiner) - Fixed issue with logging errors when a certain exception is raised in the Snapshot History function